Simplified the flow detail panel
- split into three components
  - the main panel (SelectedFlowDetailPanel)
  - the SelectedLogicalFlowDetail
  - the SelectedPhysicalFlowDetail
- bonus: fixed alignment of the RatingIndicatorCell

#6895
