+++ waltz-ng/client/ratings/components/rating-indicator-cell/rating-indicator-cell.scss
@@ -27,6 +27,7 @@
     border-radius: 6px;
 }
 
+
 .wric-name {
     position: relative;
     top: -4px;
