+++ waltz-test-common/docs/report-grid/create-grid/README.md
@@ -1,0 +1,31 @@
+# Report Grid Creation
+
+Report grids display information about whatever apps or change initiatives their parent page includes.
+In this example we are going to use starting point of a simple application group with two apps.
+We will then build up an example report pulling in some assessments, organisational and other data points.  
+
+
+
+## Open the Report Grid section
+
+Report grids are accessible from any grouping page in Waltz and can be found on the left hand navigation menu.
+
+
+| Step                                               | Image                       |
+|----------------------------------------------------|-----------------------------|
+| First we need to open the "_Report Grids_" section | ![grid_section_opened.png](2_grid_section_opened.png "grid_section_opened.png") |
+
+
+## Create the basic grid
+
+To create a grid we need to give it a name, optionally set a description and chose whether
+the grid will display information about _Applications_ or _Change Initiatives_
+
+
+| Step                                                                                                                  | Image                  |
+|-----------------------------------------------------------------------------------------------------------------------|------------------------|
+| Next, create a new grid using the "_Create a new report grid_" button                                                 | ![new_grid_form.png](4_new_grid_form.png "new_grid_form.png")  |
+| Fill in a title and description, in this example the report will be about Applications so we will leave that selected | ![completed_form.png](5_completed_form.png "completed_form.png") |
+| When we submit the form we will be taken to the column editor to customise the grid                                   | ![submitted_form.png](6_submitted_form.png "submitted_form.png") |
+
+## Customising the grid
