+++ waltz-test-common/docs/user/registration/README.md
@@ -1,0 +1,9 @@
+# User Registration
+
+
+| Step                                                                                                   | Image                                                       |
+|--------------------------------------------------------------------------------------------------------|-------------------------------------------------------------|
+| Navigate to admin page and click on user management link                                               | ![admin_page.png](1_admin_page.png "admin_page.png")                                          |
+| Select the "_add a new user_ " link                                                                    | ![add_user_button.png](2_add_user_button.png "add_user_button.png")                                     |
+| Fill in the new user registration form, once completed the "_Register_" button will be enables         | ![blank_add_user_form.png](3_blank_add_user_form.png "blank_add_user_form.png") ![completed_add_user_form.png](4_completed_add_user_form.png "completed_add_user_form.png") |
+| When you submit the form the user will be created and you will be presented with the user roles screen | ![user_registered.png](5_user_registered.png "user_registered.png")                                     |
